[Feature amity] Add testfixes for amity
