+++ .gitignore
@@ -1,0 +1,1 @@
+# Created by .ignore support plugin (hsz.mobi)
